Just in case that someone writes an exponential number, to correctly parse it.

+++ src/basic/Parser.java
@@ -393,7 +393,7 @@ public class Parser {
       numberOfSimulations = 10;
     }
     try {
-      numberOfSteps = json.getInt("numberOfSteps");
+      numberOfSteps = (int) json.getDouble("numberOfSteps");
     } catch (JSONException e) {
       numberOfSteps = -1;
     }
